+++ domainapp/example/dom/src/main/java/org/incode/domainapp/example/dom/wkt/gmap3/fixture/Gmap3DemoTearDownFixture.java
@@ -16,12 +16,12 @@ public class Gmap3DemoTearDownFixture extends FixtureScript {
     protected void execute(ExecutionContext executionContext) {
 
         if(ownedBy != null) {
-            isisJdoSupport.executeUpdate("delete from \"Gmap3ToDoItemDependencies\" where \"dependingId\" IN (select \"id\" from \"Gmap3ToDoItem\" where \"ownedBy\" = '" + ownedBy + "') ");
-            isisJdoSupport.executeUpdate("delete from \"Gmap3ToDoItemDependencies\" where \"dependentId\" IN (select \"id\" from \"Gmap3ToDoItem\" where \"ownedBy\" = '" + ownedBy + "') ");
-            isisJdoSupport.executeUpdate("delete from \"Gmap3ToDoItem\" where \"ownedBy\" = '" + ownedBy + "'");
+            isisJdoSupport.executeUpdate("delete from \"exampleWktGmap3\".\"Gmap3ToDoItemDependencies\" where \"dependingId\" IN (select \"id\" from \"exampleWktGmap3\".\"Gmap3ToDoItem\" where \"ownedBy\" = '" + ownedBy + "') ");
+            isisJdoSupport.executeUpdate("delete from \"exampleWktGmap3\".\"Gmap3ToDoItemDependencies\" where \"dependentId\" IN (select \"id\" from \"exampleWktGmap3\".\"Gmap3ToDoItem\" where \"ownedBy\" = '" + ownedBy + "') ");
+            isisJdoSupport.executeUpdate("delete from \"exampleWktGmap3\".\"Gmap3ToDoItem\" where \"ownedBy\" = '" + ownedBy + "'");
         } else {
-            isisJdoSupport.executeUpdate("delete from \"Gmap3ToDoItemDependencies\"");
-            isisJdoSupport.executeUpdate("delete from \"Gmap3ToDoItem\"");
+            isisJdoSupport.executeUpdate("delete from \"exampleWktGmap3\".\"Gmap3ToDoItemDependencies\"");
+            isisJdoSupport.executeUpdate("delete from \"exampleWktGmap3\".\"Gmap3ToDoItem\"");
         }
 
     }
